requirements.txt
alembic==1.7.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==3.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.41.2
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
